requirements.txt
google-adk
google-generativeai
mcp
faiss-cpu
fastapi
uvicorn
python-dotenv

